agent-tasty/pyproject.toml
[project]
name = "agent-tasty"
version = "0.1.0"
description = "AI conversational agent using LangGraph and DSPy"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "dspy>=2.5",
    "redis>=5.0",
    "python-dotenv>=1.0",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.2",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "faster-whisper>=1.0",
    "pymssql>=2.2",
    "requests>=2.28",
    "fpdf2>=2.7",
    "apscheduler>=3.10",
    "pytz>=2024.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
